CustomConverter.apply_filter() takes the filter itself, which contains params
