+++ simulariumio/filters/filter.py
@@ -3,9 +3,7 @@
 
 import logging
 from abc import ABC, abstractmethod
-from typing import Dict, Any
 
-from .params import FilterParams
 from ..data_objects import CustomData
 
 ###############################################################################
@@ -17,7 +15,5 @@ log = logging.getLogger(__name__)
 
 class Filter(ABC):
     @abstractmethod
-    def filter(
-        self, data: CustomData, params: FilterParams
-    ) -> CustomData:
+    def apply(self, data: CustomData) -> CustomData:
         pass
