--- simulariumio/filters/params/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from .filter_params import FilterParams  # noqa: F401
-from .every_nth_agent_filter_params import EveryNthAgentFilterParams  # noqa: F401
