--- simulariumio/filters/params/filter_params.py
@@ -1,15 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import logging
-
-###############################################################################
-
-log = logging.getLogger(__name__)
-
-###############################################################################
-
-
-class FilterParams:
-    name: str
-    pass
